Renamed repository/git.py -> git.py, and associated changes
