+++ src/mepo/command/checkout.py
@@ -1,7 +1,7 @@
 from ..state.state import MepoState
 from ..utilities import verify
 from ..utilities import colors
-from ..repository.git import GitRepository
+from ..git import GitRepository
 
 def run(args):
     allcomps = MepoState.read_state()
