+++ src/mepo/command/clone.py
@@ -7,7 +7,7 @@ from urllib.parse import urlparse
 from ..state.state import MepoState
 from ..state.state import StateDoesNotExistError
 from ..state.component import MepoComponent
-from ..repository.git import GitRepository
+from ..git import GitRepository
 from ..utilities import shellcmd
 from ..utilities import colors
 from ..utilities import mepoconfig
