+++ src/mepo/command/develop.py
@@ -1,6 +1,6 @@
 from ..state.state import MepoState
 from ..utilities import verify
-from ..repository.git import GitRepository
+from ..git import GitRepository
 from ..utilities import colors
 
 def run(args):
