+++ src/mepo/command/diff.py
@@ -6,7 +6,7 @@ import os
 from shutil import get_terminal_size
 
 from ..state.state import MepoState
-from ..repository.git import GitRepository
+from ..git import GitRepository
 from ..utilities import verify
 
 def run(args):
