+++ src/mepo/command/push.py
@@ -1,6 +1,6 @@
 from ..utilities import verify
 from ..state.state import MepoState
-from ..repository.git import GitRepository
+from ..git import GitRepository
 
 def run(args):
     allcomps = MepoState.read_state()
