+++ src/mepo/command/restore-state.py
@@ -4,7 +4,7 @@ import multiprocessing as mp
 import atexit
 
 from ..state.state import MepoState
-from ..repository.git import GitRepository
+from ..git import GitRepository
 from ..utilities.version import version_to_string
 from ..utilities import colors
