+++ src/mepo/command/stash_list.py
@@ -1,5 +1,5 @@
 from ..state.state import MepoState
-from ..repository.git import GitRepository
+from ..git import GitRepository
 
 def run(args):
     allcomps = MepoState.read_state()
